+++ src/index.ts
@@ -1,3 +1,4 @@
+export { logPackageInitialized } from "./package-init.js";
 export { resolveLogger } from "./resolve-logger.js";
 
 export type {